+++ components/form/FPhoneInput.vue
@@ -2,7 +2,7 @@
 FField.FPhoneInput(
   :class="classes"
   :style="style"
-  v-bind="{ name, label, labelTextColor, hint, hideHint, hintTextColor }"
+  v-bind="{ name, label, labelTextColor, hint, hideHint, hintTextColor, hintIcon }"
 )
   FMenu(
     :options="countries"
@@ -212,6 +212,10 @@ export interface FPhoneInputProps {
    */
   hint?: string;
   /**
+   * Icon, displayed before the hint
+   */
+  hintIcon?: Icon | null;
+  /**
    * Rules form validation
    */
   rules?: ValidationRule | ValidationRule[];
@@ -282,6 +286,7 @@ const props = withDefaults(defineProps<FPhoneInputProps>(), {
   validateOnMount: false,
   hint: '',
   hintTextColor: 'neutral--dark-4',
+  hintIcon: null,
   rules: () => [],
   errorMessage: '',
   errorColor: 'danger',