+++ components/form/FRadioGroup.vue
@@ -3,7 +3,7 @@ FField.FRadioGroup(
   fieldset
   :class="classes"
   :disabled="disabled"
-  v-bind="{ name, label, labelTextColor, hint, hideHint, hintTextColor }"
+  v-bind="{ name, label, labelTextColor, hint, hideHint, hintTextColor, hintIcon }"
 )
   .FRadioGroup__options
     FRadio(
@@ -104,6 +104,10 @@ export interface FRadioGroupProps {
    */
   hint?: string;
   /**
+   * Icon, displayed before the hint
+   */
+  hintIcon?: Icon | null;
+  /**
    * Message to use as hint when validation fails
    */
   errorMessage?: string;
@@ -119,6 +123,7 @@ const props = withDefaults(defineProps<FRadioGroupProps>(), {
   errorMessage: '',
   hint: '',
   hintTextColor: 'neutral--dark-4',
+  hintIcon: null,
   label: '',
   modelValue: '',
   name: '',